refactor(efb): unify PMDG 737 and 777 EFB into one PMDGEFBForm
Move all EFB-bridge-bound code from Forms/PMDG777/ to a new shared
Forms/PMDGEFB/ namespace and rename PMDG777EFBForm -> PMDGEFBForm. The
ctor takes the active aircraft code so the form title and AccessibleName
read "PMDG 737 EFB" or "PMDG 777 EFB" at runtime. Hard-coded title
strings are removed from the Designer and set in the ctor after
InitializeComponent().

The 738 first-cut PMDG737EFBForm (Dashboard + Preferences only) is
deleted; the 737 now reaches full parity with the 777 because the EFB
tablet app is byte-identical across all five PMDG aircraft variants
(MD5-confirmed for PMDGTablet.js across 736 / 737 / 738 / 739 / 77W).

PMDG777CDUForm stays in Forms/PMDG777/ (CDU is aircraft-specific).

MainForm.ShowPMDGEFBDialog collapses from a per-aircraft switch to a
single PMDGEFBForm construction.
